Store coordinates as decimal
Otherwise, MySQL performs an arbitrary rounding when returning results.
The precision is chosen as following:
* 3 digits before the point [-360; 360]
* 8 digits after. GeoData compares coordinates with 1e-6 precision, 2 extra digits
  to reduce the possibility of "smart" rounding.

Bug: T78715

+++ sql/externally-backed.sql
@@ -12,9 +12,9 @@ CREATE TABLE /*_*/geo_tags (
 	-- or secondary (just mentioned in text)
 	gt_primary bool NOT NULL,
 	-- Latitude of the point in degrees
-	gt_lat float NOT NULL,
+	gt_lat decimal(11,8) NOT NULL,
 	-- Longitude of the point in degrees
-	gt_lon float NOT NULL,
+	gt_lon decimal(11,8) NOT NULL,
 	-- Approximate viewing radius in meters, gives an idea how large the object is
 	gt_dim int NULL,
 	-- Type of the point
